Crunchy-Streamer: Add files to the repo.
I finished this project before creating a git repo, so I just add all files
to the repo with this commit.

I wanted to use streamlink to stream directly, but the crunchyroll plugin
is currently broken, so I am using crunchy-cli in the meantime, which
downloads the episode, then streams it to VLC, and delete the download
afterwards.


Signed-off-by: Marcel Koseler <marcelkoseler@aol.com>